+++ src/App.js
@@ -1,23 +1,13 @@
-import React, { useEffect } from 'react';
-import { useDispatch, useSelector } from 'react-redux';
-import { fetchMessages, selectMessages } from './reducers/messagesSlice';
-import './App.css';
+import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
+import Greeting from './Greeting';
 
 function App() {
-  const dispatch = useDispatch();
-  const messages = useSelector(selectMessages);
-
-  useEffect(() => {
-    dispatch(fetchMessages());
-  }, [dispatch]);
-
   return (
-    <div className="App">
-      <h1>React app</h1>
-      <h3>
-        {messages.content}
-      </h3>
-    </div>
+    <Router>
+      <Switch>
+        <Route exact path="/" component={Greeting} />
+      </Switch>
+    </Router>
   );
 }
